TrafficLaw_CHATBOT_LLM/requirements.txt
streamlit
requests
sentence-transformers
faiss-cpu
pypdf
python-docx
numpy
google-generativeai
groq
watchdog
